BuildingUpdate/ApartmentUpdate_SomeChanges

+++ src/main/java/com/buildingmanager/apartment/ApartmentDTO.java
@@ -1,52 +1,52 @@
-package com.buildingmanager.apartment;
+    package com.buildingmanager.apartment;
 
-import lombok.*;
+    import lombok.*;
 
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ApartmentDTO {
-    private Long id;
+    @Data
+    @NoArgsConstructor
+    @AllArgsConstructor
+    @Builder
+    public class ApartmentDTO {
+        private Integer id;
 
-    private String ownerFirstName;
-    private String ownerLastName;
-    private String number;
-    private String sqMetersApart;
-    private String floor;
-    private Boolean parkingSpace;
+        private String ownerFirstName;
+        private String ownerLastName;
+        private String number;
+        private String sqMetersApart;
+        private String floor;
+        private Boolean parkingSpace;
 
-    private Boolean isRented;
-    private String residentFirstName;
-    private String residentLastName;
-    private String parkingSlot;
+        private Boolean isRented;
+        private String residentFirstName;
+        private String residentLastName;
+        private String parkingSlot;
 
-    private Boolean apStorageExist;
-    private String storageSlot;
-    private Boolean isManagerHouse;
+        private Boolean apStorageExist;
+        private String storageSlot;
+        private Boolean isManagerHouse;
 
-    private Double commonPercent;
-    private Double elevatorPercent;
-    private Double heatingPercent;
+        private Double commonPercent;
+        private Double elevatorPercent;
+        private Double heatingPercent;
 
-    private Boolean active;
-    private Boolean enable;
+        private Boolean active;
+        private Boolean enable;
 
-    private String apDescription;
+        private String apDescription;
 
-    private Long buildingId;
+        private Long buildingId;
 
-    private String resident;
-    private Integer residentId;
-    private String residentEmail;
-    private String residentPhone;
+        private String resident;
+        private Integer residentId;
+        private String residentEmail;
+        private String residentPhone;
 
-    private String owner;
-    private Integer ownerId;
-    private String ownerEmail;
-    private String ownerPhone;
-    private String ownerCountry;
-    private String ownerFullName;
+        private String owner;
+        private Integer ownerId;
+        private String ownerEmail;
+        private String ownerPhone;
+        private String ownerCountry;
+        private String ownerFullName;
 
-}
+    }
